requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
Pillow>=10.0
python-decouple>=3.8
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
